Major Changes

+++ Classes/AI/MPS_ASModifier.cpp
@@ -6,7 +6,7 @@ MPS_ASModifier::MPS_ASModifier() {
 }
 MPS_ASModifier::MPS_ASModifier(Character* c) {
 	if (c->getHPCurrent() == 0) Val = 1;
-	else Val = c->getHPMax() / c->getHPCurrent();
+	else Val = (float)(c->getHPMax() / c->getHPCurrent());
 	Type = MPS_Resource::tASM;
 }
 MPS_ASModifier::MPS_ASModifier(std::string Path) {}
